requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.16.0b1
azure-storage-file-datalake==12.11.0b1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
dacite==1.8.0
ecdsa==0.18.0
fastapi==0.95.0
fastapi-azure-auth==3.5.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
isodate==0.6.1
msal==1.21.0
msal-extensions==1.0.0
multidict==6.0.4
openai==0.27.2
packaging==23.0
portalocker==2.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
replicate==0.5.2
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
